Added option to cancel a purchase Updated protocol   Changed date to ISO-Format   Useing String.format now

+++ app/src/main/java/rynkbit/tk/coffeelist/db/DbHelper.java
@@ -6,6 +6,7 @@ import android.database.sqlite.SQLiteOpenHelper;
 import android.os.Debug;
 import android.util.Log;
 
+import com.j256.ormlite.android.apptools.OrmLiteConfigUtil;
 import com.j256.ormlite.android.apptools.OrmLiteSqliteOpenHelper;
 import com.j256.ormlite.dao.Dao;
 import com.j256.ormlite.dao.DaoManager;
@@ -19,9 +20,11 @@ import java.util.List;
 import rynkbit.tk.coffeelist.BuildConfig;
 import rynkbit.tk.coffeelist.db.contract.DbContract;
 import rynkbit.tk.coffeelist.db.entity.Admin;
+import rynkbit.tk.coffeelist.db.entity.Invoice;
 import rynkbit.tk.coffeelist.db.entity.Item;
 import rynkbit.tk.coffeelist.db.entity.Protocol;
 import rynkbit.tk.coffeelist.db.entity.User;
+import rynkbit.tk.coffeelist.db.facade.InvoiceFacade;
 
 
 /**
@@ -39,10 +42,11 @@ public class DbHelper extends OrmLiteSqliteOpenHelper{
     @Override
     public void onCreate(SQLiteDatabase sqLiteDatabase, ConnectionSource connectionSource) {
         try {
-            TableUtils.createTable(connectionSource, User.class);
-            TableUtils.createTable(connectionSource, Item.class);
-            TableUtils.createTable(connectionSource, Admin.class);
-            TableUtils.createTable(connectionSource, Protocol.class);
+            TableUtils.createTableIfNotExists(connectionSource, User.class);
+            TableUtils.createTableIfNotExists(connectionSource, Item.class);
+            TableUtils.createTableIfNotExists(connectionSource, Admin.class);
+            TableUtils.createTableIfNotExists(connectionSource, Protocol.class);
+            TableUtils.createTableIfNotExists(connectionSource, Invoice.class);
 
             List<User> testUsers = new LinkedList<>();
             List<Item> testItem = new LinkedList<>();
@@ -85,16 +89,28 @@ public class DbHelper extends OrmLiteSqliteOpenHelper{
     }
 
     @Override
-    public void onUpgrade(SQLiteDatabase sqLiteDatabase, ConnectionSource connectionSource, int i, int i1) {
-        try {
-            TableUtils.dropTable(connectionSource, User.class, true);
-            TableUtils.dropTable(connectionSource, Item.class, true);
-            TableUtils.dropTable(connectionSource, Admin.class, true);
-            TableUtils.dropTable(connectionSource, Protocol.class, true);
-        } catch (SQLException e) {
-            e.printStackTrace();
+    public void onUpgrade(SQLiteDatabase sqLiteDatabase, ConnectionSource connectionSource,
+                          int oldVersion, int newVersion) {
+        if(newVersion < 15) {
+            try {
+                TableUtils.dropTable(connectionSource, User.class, true);
+                TableUtils.dropTable(connectionSource, Item.class, true);
+                TableUtils.dropTable(connectionSource, Admin.class, true);
+                TableUtils.dropTable(connectionSource, Protocol.class, true);
+            } catch (SQLException e) {
+                e.printStackTrace();
+            }
+            onCreate(sqLiteDatabase, connectionSource);
         }
-        onCreate(sqLiteDatabase, connectionSource);
+        else if(newVersion == 15 || (oldVersion < 15 && newVersion == 16)){
+            try {
+                TableUtils.createTableIfNotExists(connectionSource, Invoice.class);
+            } catch (SQLException e) {
+                e.printStackTrace();
+            }
+        }else if(oldVersion == 15 && newVersion == 16){
+            sqLiteDatabase.execSQL("alter table Invoice add column reverted integer default 0");
+        }
     }
